# perf: cache-before-normalize optimization for edge function